+++ bin/src/model/model/custom_model.dart
@@ -1,0 +1,8 @@
+import 'model.dart';
+
+class CustomModel extends Model {
+  CustomModel(
+    String name,
+    String path,
+  ) : super(name, path);
+}