+++ example/lib/model/user/testing.g.dart
@@ -12,6 +12,11 @@ Testing _$TestingFromJson(Map<String, dynamic> json) {
     beneficiary: json['beneficiary'] as String,
     structuredMessage: json['structuredMessage'] as String,
     beneficiaryIBAN: json['beneficiaryIBAN'] as String,
+    duration: json['duration'] == null
+        ? null
+        : Duration.fromJson(json['duration'] as Map<String, dynamic>),
+    durationFromJsonTest:
+        handleDurationFromToJsonFromJson(json['duration_from_json_test']),
   );
 }
 
@@ -19,4 +24,7 @@ Map<String, dynamic> _$TestingToJson(Testing instance) => <String, dynamic>{
       'beneficiary': instance.beneficiary,
       'structuredMessage': instance.structuredMessage,
       'beneficiaryIBAN': instance.beneficiaryIBAN,
+      'duration': instance.duration,
+      'duration_from_json_test':
+          handleDurationFromToJsonToJson(instance.durationFromJsonTest),
     };